feat: return function abis as part execution res
