+++ yarn-project/acir-simulator/src/simulator.ts
@@ -1,4 +1,5 @@
 import { EthAddress, OldTreeRoots, TxRequest, AztecAddress } from '@aztec/circuits.js';
+import { FunctionAbi } from '@aztec/noir-contracts';
 import { DBOracle } from './db_oracle.js';
 import { Execution, ExecutionResult } from './execution.js';
 
@@ -7,12 +8,12 @@ export class AcirSimulator {
 
   run(
     request: TxRequest,
-    entryPointACIR: Buffer,
+    entryPointABI: FunctionAbi,
     contractAddress: AztecAddress,
     portalContractAddress: EthAddress,
     oldRoots: OldTreeRoots,
   ): Promise<ExecutionResult> {
-    const execution = new Execution(this.db, request, entryPointACIR, contractAddress, portalContractAddress, oldRoots);
+    const execution = new Execution(this.db, request, entryPointABI, contractAddress, portalContractAddress, oldRoots);
 
     return execution.run();
   }
