+++ yarn-project/aztec-rpc/src/aztec_rpc_client/aztec_rpc_client.ts
@@ -2,7 +2,7 @@ import { AztecAddress, EthAddress, Fr, TxRequest } from '@aztec/circuits.js';
 import { ContractData } from '@aztec/l2-block';
 import { Tx, TxHash } from '@aztec/tx';
 import { Signature } from '../circuits.js';
-import { ContractAbi } from '../noir.js';
+import { ContractAbi } from '@aztec/noir-contracts';
 import { TxReceipt } from '../tx/index.js';
 
 export interface DeployedContract {
