+++ yarn-project/aztec-rpc/src/aztec_rpc_server/aztec_rpc_server.ts
@@ -21,7 +21,7 @@ import { Database } from '../database/database.js';
 import { TxDao } from '../database/tx_dao.js';
 import { TxReceipt, TxStatus } from '../tx/index.js';
 import { KeyStore } from '../key_store/index.js';
-import { ContractAbi, FunctionType } from '../noir.js';
+import { ContractAbi, FunctionType } from '@aztec/noir-contracts';
 import { Synchroniser } from '../synchroniser/index.js';
 
 /**
@@ -192,7 +192,7 @@ export class AztecRPCServer implements AztecRPCClient {
     const oldRoots = new OldTreeRoots(Fr.ZERO, Fr.ZERO, Fr.ZERO, Fr.ZERO); // TODO - get old roots from the database/node
     const executionResult = await this.acirSimulator.run(
       txRequest,
-      Buffer.from(functionDao.bytecode),
+      functionDao,
       contractAddress,
       contract.portalAddress,
       oldRoots,
