+++ yarn-project/aztec-rpc/src/index.ts
@@ -6,6 +6,5 @@ export * from './tx/index.js';
 export { Tx, TxHash } from '@aztec/tx';
 // TODO - only export necessary stuffs
 export * from './circuits.js';
-export * from './noir.js';
 
 export { Fr, TxRequest, AztecAddress, EthAddress } from '@aztec/circuits.js';
