+++ yarn-project/aztec.js/src/contract/contract.ts
@@ -1,4 +1,5 @@
-import { AztecAddress, AztecRPCClient, ContractAbi, FunctionAbi } from '@aztec/aztec-rpc';
+import { AztecAddress, AztecRPCClient } from '@aztec/aztec-rpc';
+import { ContractAbi, FunctionAbi } from '@aztec/noir-contracts';
 import { ContractFunctionInteraction } from './contract_function_interaction.js';
 
 export class Contract {
