+++ yarn-project/aztec.js/src/contract/contract_function_interaction.ts
@@ -1,5 +1,6 @@
-import { AztecRPCClient, Signature, Tx, TxHash, TxRequest, FunctionType } from '@aztec/aztec-rpc';
+import { AztecRPCClient, Signature, Tx, TxHash, TxRequest } from '@aztec/aztec-rpc';
 import { AztecAddress, Fr } from '@aztec/circuits.js';
+import { FunctionType } from '@aztec/noir-contracts';
 import { SentTx } from './sent_tx.js';
 
 export interface SendMethodOptions {
