+++ yarn-project/aztec.js/src/contract_deployer/contract_deployer.test.ts
@@ -1,18 +1,10 @@
 import { mock } from 'jest-mock-extended';
-import {
-  AztecRPCClient,
-  ContractAbi,
-  Signature,
-  Tx,
-  TxHash,
-  TxReceipt,
-  TxRequest,
-  FunctionType,
-} from '@aztec/aztec-rpc';
+import { AztecRPCClient, Signature, Tx, TxHash, TxReceipt, TxRequest } from '@aztec/aztec-rpc';
 
 import { ContractDeployer } from './contract_deployer.js';
 import { AztecAddress, EthAddress, Fr } from '@aztec/circuits.js';
 import { randomBytes } from 'crypto';
+import { ContractAbi, FunctionType } from '@aztec/noir-contracts';
 
 describe('Contract Deployer', () => {
   let arc: ReturnType<typeof mock<AztecRPCClient>>;
