+++ yarn-project/aztec.js/src/contract_deployer/contract_deployer.ts
@@ -1,4 +1,5 @@
-import { AztecRPCClient, ContractAbi } from '@aztec/aztec-rpc';
+import { AztecRPCClient } from '@aztec/aztec-rpc';
+import { ContractAbi } from '@aztec/noir-contracts';
 import { DeployMethod } from './deploy_method.js';
 
 /**
